requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchinfo>=1.8.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
tqdm>=4.65.0
jupyter>=1.0.0
scikit-learn>=1.2.0
opencv-python>=4.8.0
kaggle>=1.6.0
